Allow credentials access from fork PRs (#50)
Use the pull request target event instead of pull request which allows repository secrets access under the basis that it runs against the target branch so malicious PRs cannot leak the secret without already merging something to main. We then checkout the PR head which makes it equivalent to using pull_request only with access to the secrets.

The safety of this approach is based on the fact that this repository (at least at the time of writing) requires approval from someone with write access for a PR from an outside fork to run the actions. This means the this is only safe under the following conditions:

We don't change the setting allowing anyone to run actions on their fork PRs
All admins of this repo are careful which PRs they approve to run actions